fix: add test for other formats and make tests pass
